requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.2
async-timeout==5.0.1
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-allauth==65.11.2
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.17
django-redis==6.0.0
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-filter==2.1.0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
